Uzair_02_Class/pyproject.toml
[project]
name = "uzair-02-class"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "MusadiqUrRahman", email = "syedmusadiq.rahman@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents>=0.0.14",
]

[project.scripts]
uzair-02-class = "uzair_02_class:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
